+++ solr/core/src/test/org/apache/solr/handler/TestHealthCheckHandlerLegacyMode.java
@@ -17,21 +17,23 @@
 package org.apache.solr.handler;
 
 import org.apache.lucene.util.LuceneTestCase.Slow;
+import org.apache.solr.SolrTestCaseJ4;
 import org.apache.solr.SolrTestCaseJ4.SuppressSSL;
-import org.apache.solr.client.solrj.SolrRequest;
+import org.apache.solr.client.solrj.embedded.JettySolrRunner;
 import org.apache.solr.client.solrj.impl.HttpSolrClient;
-import org.apache.solr.client.solrj.request.GenericSolrRequest;
 import org.apache.solr.client.solrj.request.HealthCheckRequest;
 import org.apache.solr.common.params.CommonParams;
 import org.apache.solr.common.params.ModifiableSolrParams;
 import org.apache.solr.common.util.NamedList;
-import org.apache.solr.handler.admin.HealthCheckHandler;
 import org.apache.solr.util.TestInjection;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 
-import static org.apache.solr.common.params.CommonParams.HEALTH_CHECK_HANDLER_PATH;
+import java.io.IOException;
+
+import static org.apache.solr.handler.ReplicationTestHelper.*;
+
 
 /**
  * Test for HealthCheckHandler in legacy mode
@@ -40,120 +42,167 @@ import static org.apache.solr.common.params.CommonParams.HEALTH_CHECK_HANDLER_PA
  */
 @Slow
 @SuppressSSL     // Currently unknown why SSL does not work with this test
-public class TestHealthCheckHandlerLegacyMode extends AbstractReplicationTestBase {
+public class TestHealthCheckHandlerLegacyMode extends SolrTestCaseJ4 {
+    HttpSolrClient leaderClientHealthCheck, followerClientHealthCheck;
 
-  HttpSolrClient leaderClientHealthCheck, followerClientHealthCheck;
+    JettySolrRunner leaderJetty, followerJetty;
+    HttpSolrClient leaderClient, followerClient;
+    ReplicationTestHelper.SolrInstance leader = null, follower = null;
 
-  @Before
-  public void setUp() throws Exception {
-    super.setUp();
-    leaderClientHealthCheck = createNewSolrClient(leaderJetty.getLocalPort(), false);
-    followerClientHealthCheck = createNewSolrClient(followerJetty.getLocalPort(), false);
-  }
+    private static String context = "/solr";
 
-  public void clearIndexWithReplication() throws Exception {
-    if (numFound(query("*:*", leaderClient)) != 0) {
-      leaderClient.deleteByQuery("*:*");
-      leaderClient.commit();
-      // wait for replication to sync & verify
-      assertEquals(0, numFound(rQuery(0, "*:*", followerClient)));
+    static int nDocs = 500;
+
+    @Before
+    public void setUp() throws Exception {
+        super.setUp();
+
+        systemSetPropertySolrDisableUrlAllowList("true");
+
+        leader = new ReplicationTestHelper.SolrInstance(createTempDir("solr-instance").toFile(), "leader", null);
+        leader.setUp();
+        leaderJetty = ReplicationTestHelper.createAndStartJetty(leader);
+        leaderClient = ReplicationTestHelper.createNewSolrClient(buildUrl(leaderJetty.getLocalPort(), context) + "/" + DEFAULT_TEST_CORENAME);
+        leaderClientHealthCheck = ReplicationTestHelper.createNewSolrClient(buildUrl(leaderJetty.getLocalPort(), context));
+
+        follower = new SolrInstance(createTempDir("solr-instance").toFile(), "follower", leaderJetty.getLocalPort());
+        follower.setUp();
+        followerJetty = createAndStartJetty(follower);
+        followerClient = ReplicationTestHelper.createNewSolrClient(buildUrl(followerJetty.getLocalPort(), context) + "/" + DEFAULT_TEST_CORENAME);
+        followerClientHealthCheck = ReplicationTestHelper.createNewSolrClient(buildUrl(followerJetty.getLocalPort(), context));
+
+        System.setProperty("solr.indexfetcher.sotimeout2", "45000");
     }
-  }
 
-  @Override
-  @After
-  public void tearDown() throws Exception {
-    super.tearDown();
-
-    if(null != leaderClientHealthCheck)  {
-      leaderClientHealthCheck.close();
-      leaderClientHealthCheck = null;
+    public void clearIndexWithReplication() throws Exception {
+        if (numFound(ReplicationTestHelper.query("*:*", leaderClient)) != 0) {
+            leaderClient.deleteByQuery("*:*");
+            leaderClient.commit();
+            // wait for replication to sync & verify
+            assertEquals(0, numFound(rQuery(0, "*:*", followerClient)));
+        }
     }
 
-    if(null != followerClientHealthCheck)  {
-      followerClientHealthCheck.close();
-      followerClientHealthCheck = null;
+    @Override
+    @After
+    public void tearDown() throws Exception {
+        super.tearDown();
+        if (null != leaderJetty) {
+            leaderJetty.stop();
+            leaderJetty = null;
+        }
+        if (null != followerJetty) {
+            followerJetty.stop();
+            followerJetty = null;
+        }
+        if (null != leaderClient) {
+            leaderClient.close();
+            leaderClient = null;
+        }
+        if (null != followerClient) {
+            followerClient.close();
+            followerClient = null;
+        }
+        if (null != leaderClientHealthCheck) {
+            leaderClientHealthCheck.close();
+            leaderClientHealthCheck = null;
+        }
+
+        if (null != followerClientHealthCheck) {
+            followerClientHealthCheck.close();
+            followerClientHealthCheck = null;
+        }
+        System.clearProperty("solr.indexfetcher.sotimeout");
     }
-  }
 
-  @Test
-  // keep this
-  public void doTestHealthCheckWithReplication() throws Exception {
 
-    TestInjection.delayBeforeFollowerCommitRefresh = random().nextInt(10);
+    @Test
+    // keep this
+    public void doTestHealthCheckWithReplication() throws Exception {
 
-    //clearIndexWithReplication();
+        TestInjection.delayBeforeFollowerCommitRefresh = random().nextInt(10);
 
-    // stop replication so that the follower doesn't pull the index
-    invokeReplicationCommand(followerJetty.getLocalPort(), "disablepoll");
+        // stop replication so that the follower doesn't pull the index
+        invokeReplicationCommand(buildUrl(followerJetty.getLocalPort(), context) + "/" + DEFAULT_TEST_CORENAME, "disablepoll");
 
-    nDocs--;
-    // create multiple commits
-    int docsAdded = 0;
-    for (int i = 0; docsAdded < nDocs / 2; i++, docsAdded++) {
-      index(leaderClient, "id", i, "name", "name = " + i);
-      if( i % 10 == 0) {
+        nDocs--;
+        // create multiple commits
+        int docsAdded = 0;
+        for (int i = 0; docsAdded < nDocs / 2; i++, docsAdded++) {
+            index(leaderClient, "id", i, "name", "name = " + i);
+            if (i % 10 == 0) {
+                leaderClient.commit();
+            }
+        }
+
         leaderClient.commit();
-      }
-    }
 
-    leaderClient.commit();
+        assertNumFoundWithQuery(leaderClient, docsAdded);
 
-    assertNumFoundWithQuery(leaderClient, docsAdded);
+        // ensure that the leader is always happy
+        // first try without specifying maxGenerationLag lag
+        ModifiableSolrParams solrParamsLeaderHealthCheck = new ModifiableSolrParams();
+        HealthCheckRequest healthCheckRequestLeader = new HealthCheckRequest();
+        assertEquals(CommonParams.OK, healthCheckRequestLeader.process(leaderClientHealthCheck).getResponse().get(CommonParams.STATUS));
 
-    //ensure that the leader is always happy
-    //first try without specifying maxGenerationLag lag
-    ModifiableSolrParams solrParamsLeaderHealthCheck = new ModifiableSolrParams();
-    SolrRequest healthCheckRequestLeader = new GenericSolrRequest(SolrRequest.METHOD.GET, HEALTH_CHECK_HANDLER_PATH, solrParamsLeaderHealthCheck);
-    assertEquals(CommonParams.OK, healthCheckRequestLeader.process(leaderClientHealthCheck).getResponse().get(CommonParams.STATUS));
+        // now try adding maxGenerationLag request param
+        solrParamsLeaderHealthCheck.add(HealthCheckRequest.PARAM_MAX_GENERATION_LAG, "2");
+        assertEquals(CommonParams.OK, healthCheckRequestLeader.process(leaderClientHealthCheck).getResponse().get(CommonParams.STATUS));
 
-    // now try adding maxGenerationLag request param
-    solrParamsLeaderHealthCheck.add(HealthCheckHandler.PARAM_MAX_GENERATION_LAG, "2");
-    assertEquals(CommonParams.OK, healthCheckRequestLeader.process(leaderClientHealthCheck).getResponse().get(CommonParams.STATUS));
+        // follower should report healthy if maxGenerationLag is not specified
+        HealthCheckRequest healthCheckRequestFollower = new HealthCheckRequest();
+        assertEquals(CommonParams.OK, healthCheckRequestFollower.process(followerClientHealthCheck).getResponse().get(CommonParams.STATUS));
 
-    //follower should report healthy if maxGenerationLag is not specified
-    SolrRequest healthCheckRequestFollower = new HealthCheckRequest();
-            new GenericSolrRequest(SolrRequest.METHOD.GET, HEALTH_CHECK_HANDLER_PATH, new ModifiableSolrParams());
-    assertEquals(CommonParams.OK, healthCheckRequestFollower.process(followerClientHealthCheck).getResponse().get(CommonParams.STATUS));
+        // ensure follower is unhealthy when maxGenerationLag is specified
+        // ModifiableSolrParams params = new ModifiableSolrParams();
+        healthCheckRequestFollower = new HealthCheckRequest();
+        healthCheckRequestFollower.setMaxGenerationLag(2);
+        assertEquals(CommonParams.FAILURE, healthCheckRequestFollower.process(followerClientHealthCheck).getResponse().get(CommonParams.STATUS));
 
-    //ensure follower is unhealthy when maxGenerationLag is specified
-    ModifiableSolrParams params = new ModifiableSolrParams();
-    params.add(HealthCheckHandler.PARAM_MAX_GENERATION_LAG, "2");
-    healthCheckRequestFollower = new GenericSolrRequest(SolrRequest.METHOD.GET, HEALTH_CHECK_HANDLER_PATH, params);
-    assertEquals(CommonParams.FAILURE, healthCheckRequestFollower.process(followerClientHealthCheck).getResponse().get(CommonParams.STATUS));
+        // enable polling, force replication and ensure that the follower is healthy
+        //invokeReplicationCommand(buildUrl(followerJetty.getLocalPort(), context) + "/" + DEFAULT_TEST_CORENAME, "enablepoll");
+        pullFromTo(leaderJetty, followerJetty);
 
-    //enable polling, force replication and ensure that the follower is healthy
-    pullFromLeaderToFollower();
+        // check replicated and is healthy
+        assertNumFoundWithQuery(followerClient, docsAdded);
+        assertEquals(CommonParams.OK, healthCheckRequestFollower.process(followerClientHealthCheck).getResponse().get(CommonParams.STATUS));
+        assertVersions(leaderClient, followerClient);
 
-    // check replicated and is healthy
-    assertNumFoundWithQuery(followerClient, docsAdded);
-    assertEquals(CommonParams.OK, healthCheckRequestFollower.process(followerClientHealthCheck).getResponse().get(CommonParams.STATUS));
-    assertVersions(leaderClient, followerClient);
-
-    //index more docs on the leader
-    for (int i = docsAdded; docsAdded < nDocs; i++, docsAdded++) {
-      index(leaderClient, "id", i, "name", "name = " + i);
-      if( i % 10 == 0) {
+        // index more docs on the leader
+        for (int i = docsAdded; docsAdded < nDocs; i++, docsAdded++) {
+            index(leaderClient, "id", i, "name", "name = " + i);
+            if (i % 10 == 0) {
+                leaderClient.commit();
+            }
+        }
         leaderClient.commit();
-      }
+
+        assertNumFoundWithQuery(leaderClient, docsAdded);
+
+        // we have added docs to the leader and polling is disabled, this should fail
+        healthCheckRequestFollower = new HealthCheckRequest();
+        healthCheckRequestFollower.setMaxGenerationLag(2);
+        assertEquals(CommonParams.FAILURE, healthCheckRequestFollower.process(followerClientHealthCheck).getResponse().get(CommonParams.STATUS));
+
+        //force replication and ensure that the follower is healthy
+        pullFromTo(leaderJetty, followerJetty);
+        assertEquals(CommonParams.OK, healthCheckRequestFollower.process(followerClientHealthCheck).getResponse().get(CommonParams.STATUS));
+        assertNumFoundWithQuery(followerClient, docsAdded);
+        assertVersions(leaderClient, followerClient);
     }
-    leaderClient.commit();
 
-    assertNumFoundWithQuery(leaderClient, docsAdded);
+    public static void pullFromTo(JettySolrRunner srcSolr, JettySolrRunner destSolr) throws IOException {
+        String srcUrl = buildUrl(srcSolr.getLocalPort(), context) + "/" + DEFAULT_TEST_CORENAME;
+        String destUrl = buildUrl(destSolr.getLocalPort(), context) + "/" + DEFAULT_TEST_CORENAME;
+        ReplicationTestHelper.pullFromTo(srcUrl, destUrl);
+    }
 
-    // we have added docs to the leader and polling is disabled, this should fail
-    healthCheckRequestFollower = new GenericSolrRequest(SolrRequest.METHOD.GET, HEALTH_CHECK_HANDLER_PATH, params);
-    assertEquals(CommonParams.FAILURE, healthCheckRequestFollower.process(followerClientHealthCheck).getResponse().get(CommonParams.STATUS));
+    private void assertNumFoundWithQuery(HttpSolrClient client, int nDocs) throws Exception {
+        NamedList<Object> queryRsp = rQuery(nDocs, "*:*", client);
+        assertEquals(nDocs, numFound(queryRsp));
+    }
 
-    //force replication and ensure that the follower is healthy
-    pullFromLeaderToFollower();
-    assertEquals(CommonParams.OK, healthCheckRequestFollower.process(followerClientHealthCheck).getResponse().get(CommonParams.STATUS));
-    assertNumFoundWithQuery(followerClient, docsAdded);
-    assertVersions(leaderClient, followerClient);
-  }
-
-  private void assertNumFoundWithQuery(HttpSolrClient client, int nDocs) throws Exception {
-    NamedList queryRsp = rQuery(nDocs, "*:*", client);
-    assertEquals(nDocs, numFound(queryRsp));
-  }
+    private static String buildUrlForCore(final int port, final String core) {
+        return buildUrl(port, context) + "/" + DEFAULT_TEST_CORENAME;
+    }
 }
